0.5.3 more consistency in types

+++ include/ffsm2/detail/shared/array.hpp
@@ -3,7 +3,7 @@ namespace detail {
 
 ////////////////////////////////////////////////////////////////////////////////
 
-template <typename T, unsigned NCapacity>
+template <typename T, Long NCapacity>
 class StaticArrayT {
 public:
 	using Item  = T;
@@ -51,7 +51,7 @@ struct StaticArrayT<T, 0> {
 
 //------------------------------------------------------------------------------
 
-template <typename T, unsigned NCapacity>
+template <typename T, Long NCapacity>
 class ArrayT {
 	template <typename>
 	friend class IteratorT;
@@ -82,7 +82,7 @@ public:
 	FFSM2_INLINE Index count()									const noexcept	{ return _count;	}
 
 	FFSM2_INLINE ArrayT& operator += (const Item& item)				  noexcept;
-	FFSM2_INLINE ArrayT& operator += (	  Item&& item)				  noexcept;
+	FFSM2_INLINE ArrayT& operator += (	   Item&& item)				  noexcept;
 
 	template <Long N>
 	FFSM2_INLINE ArrayT& operator += (const ArrayT<Item, N>& other)	  noexcept;
